add-categoryController-and-update-DTOs
