+++ ValhallaVaultCyberAwareness/API/DataTransferObjects/CategoryDTO.cs
@@ -2,6 +2,7 @@
 {
     public class CategoryDTO
     {
+        public int Id { get; set; }
         public string Title { get; set; } = null!;
         public string? Info { get; set; }
     }
