+++ ValhallaVaultCyberAwareness/API/DataTransferObjects/CategoryPostDTO.cs
@@ -1,9 +1,8 @@
 ﻿namespace ValhallaVaultCyberAwareness.API.DataTransferObjects
 {
-    public class SegmentDTO
+    public class CategoryPostDTO
     {
-        public string Name { get; set; } = null!;
+        public string Title { get; set; } = null!;
         public string? Info { get; set; }
-        public int CategoryId { get; set; }
     }
 }
