+++ ValhallaVaultCyberAwareness/API/ValhallaController.cs
@@ -26,90 +26,7 @@ namespace ValhallaVaultCyberAwareness.API
 
         #region HttpGetAll
 
-        [HttpGet("GetAllCategoriesAndMetadata")]
-        public async Task<IActionResult> GetAllCategoriesAndMetdata()
-        {
-            //Get all categories
-            var Cate = await uow.CategoryRepo.GetAllAsync(c => c.Segments);
-            var Sub = await uow.SubcategoryRepo.GetAllAsync();
-            var Ques = await uow.QuestionRepo.GetAllAsync(q => q.Answers, q => q.UserResponse);
 
-            if (Cate.Any())
-            {
-                //Turn Dbmodel to Apimodel<
-                var ApiCategoriesToReturn = Cate.Select(q => new CategoryApiModel
-                {
-                    Id = q.Id,
-                    Name = q.Name,
-                    Segments = q.Segments.Select(s => new SegmentApiModel
-                    {
-                        Id = s.Id,
-                        Name = s.Name,
-                        CategoryId = s.CategoryId,
-                        Subcategorys = s.Subcategorys.Where(su => su.Segment.Id == s.Id).Select(sc => new SubcategoryApiModel
-                        {
-                            Id = sc.Id,
-                            Name = sc.Name,
-                            SegmentId = sc.SegmentId,
-                            Questions = Ques.Where(q => q.SubcategoryId == sc.Id).Select(q => new QuestionApiModel
-                            {
-                                Title = q.Title,
-                                SubcategoryId = q.SubcategoryId,
-                                Answers = q.Answers.Select(a => new AnswerApiModel
-                                {
-                                    Id = a.Id,
-                                    Answer = a.Answer,
-                                    IsCorrectAnswer = a.IsCorrectAnswer,
-                                    Explanation = a.Explanation,
-                                    QuestionId = a.QuestionId
-                                }).ToList(),
-
-                            }).ToList()
-                        }).ToList()
-                    }).ToList()
-
-                }).ToList();
-
-
-
-
-
-                return Ok(ApiCategoriesToReturn);
-
-            }
-            else
-            {
-                return NotFound("No categories available");
-            }
-
-        }
-
-        [HttpGet("GetAllCategories")]
-        public async Task<IActionResult> GetAllCategories()
-        {
-            //Get all categories
-            var Cate = await uow.CategoryRepo.GetAllAsync(c => c.Segments);
-
-
-            if (Cate.Any())
-            {
-                //Turn Dbmodel to Apimodel<
-                var ApiCategoriesToReturn = Cate.Select(q => new CategoryApiModel
-                {
-                    Id = q.Id,
-                    Name = q.Name
-
-                }).ToList();
-
-
-                return Ok(ApiCategoriesToReturn);
-            }
-            else
-            {
-                return NotFound("No categories available");
-            }
-
-        }
 
 
         [HttpGet("GetAllSegmentsAndSubcategories")]
@@ -354,34 +271,6 @@ namespace ValhallaVaultCyberAwareness.API
         //}
 
 
-
-
-        [HttpGet("GetCategoryById/{id}")]
-        public async Task<IActionResult> GetCategoryByIdAsync(int id)
-        {
-            var Cat = await uow.CategoryRepo.GetByIdAsync(id);
-
-            if (Cat != null)
-            {
-
-                //Turn Dbmodel to Apimodel<
-                var ApiCategoryToReturn = new CategoryApiModel
-                {
-                    Id = Cat.Id,
-                    Name = Cat.Name,
-                };
-
-
-                return Ok(ApiCategoryToReturn);
-
-            }
-            else
-            {
-                return NotFound("No category found  with that id!");
-            }
-
-        }
-
         [HttpGet("GetSegmentById/{id}")]
         public async Task<IActionResult> GetSegmentByIdAsync(int id)
         {
@@ -546,29 +435,29 @@ namespace ValhallaVaultCyberAwareness.API
 
         #region HttpPost
 
-        [HttpPost("OnPostCategory")]
-        //Category
-        public async Task<IActionResult> OnPostCategory([FromBody] CategoryDTO category)
-        {
-            if (category == null)
-            {
-                return BadRequest("Error: Invalid input! Please try again..");
-            }
-            else
-            {
-                // Transfer the DTO to Db-Model
-                CategoryModel categoryToAdd = new()
-                {
-                    Name = category.Title,
-                    Info = category.Info
-                };
+        //[HttpPost("OnPostCategory")]
+        ////Category
+        //public async Task<IActionResult> OnPostCategory([FromBody] CategoryDTO category)
+        //{
+        //    if (category == null)
+        //    {
+        //        return BadRequest("Error: Invalid input! Please try again..");
+        //    }
+        //    else
+        //    {
+        //        // Transfer the DTO to Db-Model
+        //        CategoryModel categoryToAdd = new()
+        //        {
+        //            Name = category.Title,
+        //            Info = category.Info
+        //        };
 
-                await uow.CategoryRepo.CreateAsync(categoryToAdd);
-                await uow.SaveChanges();
+        //        await uow.CategoryRepo.CreateAsync(categoryToAdd);
+        //        await uow.SaveChanges();
 
-                return Ok("Category was successfully added!");
-            }
-        }
+        //        return Ok("Category was successfully added!");
+        //    }
+        //}
 
         [HttpPost("OnPostSegment")]
         //Segment
@@ -667,46 +556,7 @@ namespace ValhallaVaultCyberAwareness.API
 
         #endregion
 
-
-        #region HttpPutById
-
-        /*[HttpPut("id")]
-        public async Task<List<UserApiModel>> GetUsByIdAsync(int id)
-        {
-            return new List<UserApiModel>();
-        }*/
-
-
-        #endregion
-
-
         #region HttpDeleteById
-
-        [HttpDelete("DeleteCategory/{id}")]
-        //Category
-        public async Task<IActionResult> DeleteCategory(int id)
-        {
-            if (id <= 0)
-            {
-                return BadRequest("The Id can't be less than 1");
-            }
-            else
-            {
-                try
-                {
-                    //This variable is only here to check if it exists, if not, it throws exception.
-                    var category = await uow.CategoryRepo.GetByIdAsync(id);
-                    await uow.CategoryRepo.DeleteAsync(id);
-                    await uow.SaveChanges();
-
-                    return Ok("Category was successfully deleted!");
-                }
-                catch (InvalidOperationException)
-                {
-                    return NotFound("There is no category with that id! Please try again...");
-                }
-            }
-        }
 
         [HttpDelete("DeleteSegment/{id}")]
         //Segment
